backend/requirements.txt
Flask==3.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
opencv-python-headless==4.9.0.80
Pillow==10.4.0
PyMySQL==1.1.0
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.2
tensorflow==2.16.1
